requirements.txt
Click==7.0
dnspython==1.16.0
Flask==1.1.1
gunicorn==20.0.0
itsdangerous==1.1.0
Jinja2==2.10.3
MarkupSafe==1.1.1
pymongo==3.7.2
python-dotenv==0.10.3
Werkzeug==0.16.0
